copy-pasta'd a bunch from https://github.com/bolinfest/monaco-tm to get TextMate grammars working.
For the moment, hardcoded editors to use be `defaultLanguage="python"`
because that's what was handy in
https://github.com/bolinfest/monaco-tm. Forgot that a major
challenge was that Monaco comes with its own set of language
configurations by default. This used to be a pain to workaround,
but now https://yarnpkg.com/package/monaco-editor-core makes this
quite a bit easier, so I was able to use that with loader.config()
to avoid pulling in the default language configs.

The end result is that Python code in the editor is syntax highlighted
exactly as it appears in desktop VS Code, so things are definitely
working! Perhaps it is time to polish https://github.com/bolinfest/monaco-tm
and make it easier to reuse.